requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-aws>=0.2.0

# Document Processing
pymupdf4llm>=0.0.17
pymupdf>=1.24.0
Pillow>=10.0.0

# AWS Services
boto3>=1.35.0
botocore>=1.35.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sqlalchemy>=2.0.0

# Embeddings & Vector Operations
numpy>=1.24.0
tiktoken>=0.7.0

# MCP Server & API
fastmcp>=0.2.0
pydantic>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiofiles>=23.2.1
httpx>=0.26.0
sse-starlette>=1.8.2

# Cohere API
cohere>=5.0.0

# UI Framework
streamlit>=1.38.0
sseclient-py>=1.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Text Splitting
langchain-text-splitters>=0.3.0

# JSON handling
jsonlines>=4.0.0

# Progress tracking
tqdm>=4.66.0
